requirements.txt
google-generativeai
chromadb
pytesseract
pillow
pandas
tqdm
python-dotenv